[REFACTOR] Consolidate Game Options and Preferences
- Rename "Appearance & Behaviour" to "Preferences"
- Remove standalone Notifications page
- Update menu structure and translations
- Remove notification support check
