+++ src/features/island/hud/components/settings-menu/general-settings/BehaviourSettings.tsx
@@ -13,20 +13,22 @@ export const BehaviourSettings: React.FC = () => {
   } = useContext(Context);
 
   return (
-    <div className="flex flex-col items-start gap-2">
-      <h1 className="text-lg mb-2">
-        {t("gameOptions.generalSettings.behaviour")}
-      </h1>
-      <Switch
-        checked={showAnimations}
-        onChange={toggleAnimations}
-        label={"Animations"}
-      />
-      <Switch
-        checked={enableQuickSelect}
-        onChange={toggleQuickSelect}
-        label={"Quick Select"}
-      />
-    </div>
+    <>
+      <div className="flex flex-col items-start gap-2">
+        <h1 className="text-lg mb-2">
+          {t("gameOptions.generalSettings.behaviour")}
+        </h1>
+        <Switch
+          checked={showAnimations}
+          onChange={toggleAnimations}
+          label={"Animations"}
+        />
+        <Switch
+          checked={enableQuickSelect}
+          onChange={toggleQuickSelect}
+          label={"Quick Select"}
+        />
+      </div>
+    </>
   );
 };
