+++ src/lib/i18n/dictionaries/dictionary.json
@@ -1784,6 +1784,7 @@
   "gameOptions.blockchainSettings.swapPOLForSFL": "Swap POL for SFL",
   "gameOptions.blockchainSettings.transferOwnership": "Transfer Ownership",
   "gameOptions.generalSettings": "General Settings",
+  "gameOptions.generalSettings.notifications": "Notifications",
   "gameOptions.generalSettings.connectDiscord": "Connect to Discord",
   "gameOptions.generalSettings.assignRole": "Assign Discord Role",
   "gameOptions.generalSettings.changeLanguage": "Change Language",
@@ -1797,7 +1798,7 @@
   "gameOptions.generalSettings.disableQuickSelect": "Disable QuickSelect",
   "gameOptions.generalSettings.enableQuickSelect": "Enable QuickSelect",
   "gameOptions.generalSettings.share": "Share",
-  "gameOptions.generalSettings.appearance&behaviour": "Appearance & Behaviour",
+  "gameOptions.generalSettings.preferences": "Preferences",
   "gameOptions.plazaSettings": "Plaza Settings",
   "gameOptions.plazaSettings.title.mutedPlayers": "Muted Players",
   "gameOptions.plazaSettings.title.keybinds": "Keybinds",
